+++ srcs/requirements/bonus/clay/tools/init_setup.sh
@@ -1,8 +1,14 @@
 #!/bin/sh
 
-clay new myapp
+if [ ! -f "/clayweb_created" ]
+then
+	clay new myapp
 
-cp index.html myapp/
+	cp index.html myapp/
+
+	touch /clayweb_created
+	chmod 444 /clayweb_created
+fi
 
 cd myapp
